% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_raster.R
\name{add_area_km2_to_spatial_df}
\alias{add_area_km2_to_spatial_df}
\title{Add area in km^2 to data frame with spatial information (lon/lat or X/Y)}
\usage{
add_area_km2_to_spatial_df(df)
}
\arguments{
\item{df}{data frame with spatial information (lon/lat or X/Y)}
}
\value{

}
\description{
Add area in km^2 to data frame with spatial information (lon/lat or X/Y)
}
