% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_maps.R
\name{add_points_to_maps}
\alias{add_points_to_maps}
\title{Add points to the existing map.}
\usage{
add_points_to_maps(
  p,
  df_point,
  col_name_to_plot,
  col_name_for_size = NULL,
  color_or_fill = "fill",
  pt_size = 1.8,
  pt_stroke = 0.5
)
}
\arguments{
\item{p}{ggplot2 object}

\item{df_point}{data frame that contains the point data to plot}

\item{col_name_to_plot}{the column name of df_point for 'fill' or 'color'}

\item{col_name_for_size}{the column name of df_point for size}

\item{color_or_fill}{one of "fill" and "color"}

\item{pt_size}{the size of the point to be plotted}

\item{pt_stroke}{the stroke (the line thickness of the outline of the shape)}
}
\value{
p a ggplot2 object with points added
}
\description{
Add points to the existing map.
}
\examples{

}
