% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_maps.R
\name{add_state_to_df}
\alias{add_state_to_df}
\title{Add state name to the data frame}
\usage{
add_state_to_df(df, key_col)
}
\arguments{
\item{key_col}{this column is used to join the df_sf to the original df}
}
\value{

}
\description{
Add state name to the data frame
}
