% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_raster.R
\name{check_if_regular_grid}
\alias{check_if_regular_grid}
\title{Check if the data frame is on a regular (regularly spaced in X/Y) grid}
\usage{
check_if_regular_grid(df, res, digits = 5)
}
\arguments{
\item{df}{data frame with X/Y or lon/lat information}

\item{res}{the spatial resolution of df}

\item{digits}{(from rasterFromXYZ) numeric, indicating the requested precision for detecting whether points are on a regular grid (a low number of digits is a low precision)}
}
\value{
TRUE if on a regular grid
}
\description{
Check if the data frame is on a regular (regularly spaced in X/Y) grid
}
\note{
Adapted from raster::rasterFromXYZ
}
