% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_raster.R
\name{convert_df_to_sf_obj}
\alias{convert_df_to_sf_obj}
\title{Convert data frame to sf object}
\usage{
convert_df_to_sf_obj(df, crs_num_or_str = 4326)
}
\arguments{
\item{df}{data frame with lon and lat}

\item{crs_num_or_str}{projection string or the corresponding EPSG number}
}
\value{
sf object.
}
\description{
Convert data frame to sf object
}
\examples{
convert_df_to_sf_obj (df, 4326)

}
