% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_raster.R
\name{convert_df_xyz_to_raster}
\alias{convert_df_xyz_to_raster}
\title{Convert data frame with xyz information to raster object.}
\usage{
convert_df_xyz_to_raster(df, col_to_rasterize)
}
\arguments{
\item{df}{data frame with x, y, z information}

\item{col_to_rasterize}{the column name for "z"}
}
\value{

}
\description{
Convert data frame with xyz information to raster object.
}
