% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_raster.R
\name{crop_sf_obj_by_polygon_mask}
\alias{crop_sf_obj_by_polygon_mask}
\title{Crop sf obj by polygon}
\usage{
crop_sf_obj_by_polygon_mask(sf_obj, polygon_mask)
}
\arguments{
\item{sf_obj}{sf object to be cropped}

\item{polygon_mask}{polygon mask}
}
\value{
cropped sf object
}
\description{
Crop sf obj by polygon
}
\examples{
crop_sf_obj_by_polygon_mask (sf_obj, polygon_mask)
}
