% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_raster.R
\name{crop_sf_obj_by_rect_bb}
\alias{crop_sf_obj_by_rect_bb}
\title{Crop sf object by rectangle}
\usage{
crop_sf_obj_by_rect_bb(sf_obj, xmin, xmax, ymin, ymax)
}
\arguments{
\item{sf_obj}{sf object to be cropped}

\item{xmin}{x (or lon) min}

\item{xmax}{x (or lon) max}

\item{ymin}{y (or lat) min}

\item{ymax}{y (or lat) max}
}
\value{
cropped sf object
}
\description{
Crop sf object by rectangle
}
