% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_raster.R
\name{df_with_lonlat_to_raster}
\alias{df_with_lonlat_to_raster}
\title{Convert data frame with lon/lat to a raster object}
\usage{
df_with_lonlat_to_raster(df)
}
\arguments{
\item{df}{data frame with columns of lon ant lat}
}
\value{
a raster object
}
\description{
Convert data frame with lon/lat to a raster object
}
\examples{
essentially same as convert_df_xyz_to_raster
}
