% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_raster.R
\name{get_crs_from_sf_obj}
\alias{get_crs_from_sf_obj}
\title{Get projection info from a sf object.}
\usage{
get_crs_from_sf_obj(sf_obj)
}
\arguments{
\item{sf_obj}{object to extract projection info from.}
}
\value{
projection info.
}
\description{
Get projection info from a sf object.
}
\examples{
get_crs_from_sf_obj (sf_obj)
}
