% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_spatial.R
\name{get_ncdf_var}
\alias{get_ncdf_var}
\title{Get a variable from netcdf file or ncdf4 object.}
\usage{
get_ncdf_var(ncin, var_name)
}
\arguments{
\item{ncin}{string or ncdf4 object}

\item{var_name}{variable name of type string.}
}
\value{

}
\description{
Get a variable from netcdf file or ncdf4 object.
}
\examples{
get_ncdf_var (ncin, var_name)
}
