% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_projection.R
\name{get_projection_string}
\alias{get_projection_string}
\title{Get projection string}
\usage{
get_projection_string(code = 4326)
}
\arguments{
\item{code}{integer for EPSG code numbers}
}
\value{

}
\description{
Get projection string
}
