% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_map.R
\name{get_us_zipcode}
\alias{get_us_zipcode}
\title{Get US Zip Code boundaries in sf object}
\usage{
get_us_zipcode(epsg_num = 4326)
}
\arguments{
\item{epsg_num}{}
}
\value{

}
\description{
Get US Zip Code boundaries in sf object
}
