% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_raster.R
\name{make_raster_template}
\alias{make_raster_template}
\title{Make raster template for regridding}
\usage{
make_raster_template(x_min, x_max, y_min, y_max, res)
}
\arguments{
\item{x_min}{the minimum X or lon (not center of the grid cell)}

\item{x_max}{the maximum X or lon (not center of the grid cell)}

\item{y_min}{the minimum Y or lat (not center of the grid cell)}

\item{y_max}{the maximum Y or lat (not center of the grid cell)}

\item{res}{spatial resolution in degrees. Assumes X and Y should have the same resolution.}
}
\value{
make raster template for regridding for any grids
}
\description{
Make raster template for regridding
}
