% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_spatial.R
\name{matrix_to_xyz_df}
\alias{matrix_to_xyz_df}
\title{Convert 2-d matrix with no column or row names to a data frame.}
\usage{
matrix_to_xyz_df(mat, var_1_name = "x", var_2_name = "y")
}
\arguments{
\item{mat}{2-d matrix with no column or row names}

\item{var_1_name}{var_1_name variable name for the first dim, default to 'x'}

\item{var_2_name}{var_2_name variable name for the second dim, default to 'y'}
}
\value{
a data frame with xyz values
}
\description{
Convert 2-d matrix with no column or row names to a data frame.
}
\examples{
foo = matrix_to_xyz_df (volcano)
}
