% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_raster.R
\name{raster_to_df}
\alias{raster_to_df}
\title{Convert raster to data frame}
\usage{
raster_to_df(r, name_4_input_val, name_4_out_val = "val")
}
\arguments{
\item{r}{input raster}

\item{name_4_input_val}{the layer name in the input raster}

\item{name_4_out_val}{the column name in the output data frame}
}
\value{

}
\description{
Convert raster to data frame
}
