% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_raster.R
\name{regrid_using_resample}
\alias{regrid_using_resample}
\title{Regrid raster object using the resample function of the raster package}
\usage{
regrid_using_resample(r, r_new_grid, method = "ngb")
}
\arguments{
\item{r}{Raster* object to be regridded}

\item{r_new_grid}{Raster* object with parameters that r should be resampled to}

\item{method}{sampling method. If the two (in & out) grids are similar in resolution, use 'ngb'. The other option is 'bilinear'}
}
\value{

}
\description{
Regrid raster object using the resample function of the raster package
}
