% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{rename_factor}
\alias{rename_factor}
\title{Rename factors}
\usage{
rename_factor(x, vector_map)
}
\arguments{
\item{x:}{factor values to be renamed}

\item{vector_map:}{one to one map for the factors with renamed factors}
}
\value{
renamed factors
}
\description{
Rename factors
}
\examples{
rename_factor (x, vector_map)
vector_map = c('< 3.0' = '727', '3.0 to 3.9' = '702', '4.0 to 4.9' = '612',
}
