% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_spatial.R
\name{list_to_pivot_longer}
\alias{list_to_pivot_longer}
\title{Make a long df from a list}
\usage{
list_to_pivot_longer(li, names_of_list, col_name_to_add)
}
\arguments{
\item{li}{a list with different lengths for each element of the list}

\item{name_for_list_names}{name for the names of the (original) list}

\item{name_for_value}{the column name for added values}
}
\value{

}
\description{
Make a long df from a list
}
