% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_raster.R
\name{plot_raster_layer_spatial}
\alias{plot_raster_layer_spatial}
\title{Plot raster object using ggspatial}
\usage{
plot_raster_layer_spatial(r)
}
\arguments{
\item{r}{raster object}
}
\value{

}
\description{
Plot raster object using ggspatial
}
\examples{

}
